% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSE.R, R/package.R
\docType{package}
\name{CMSE}
\alias{CMSE}
\title{CMSE: Compute the Causal Mean Squared Error}
\usage{
CMSE(
  dataset,
  intervention,
  models,
  posttest,
  outcomes,
  covariates = NULL,
  ...,
  nrows = NA,
  latentPosttest = NA
)
}
\arguments{
\item{dataset}{the actual data as a data frame or mxDataStatic object}

\item{intervention}{character string containing the name of the intervention column}

\item{models}{a list of mxModel objects (like those returned from makeARPanelModels) for fitting}

\item{posttest}{character string containing the name of the posttest outcome}

\item{outcomes}{character vector including names of all additional outcomes}

\item{covariates}{character vector including the names of all covariates to be included}

\item{...}{Does not accept arguments; only there so that later arguments must be named}

\item{nrows}{number of data rows in the data set; Required if dataset is a covariance matrix without means}

\item{latentPosttest}{the latent true posttest score; Required if the posttest outcome has a measurement model (e.g. for the RI-AR model)}
}
\value{
a list containing elements:
        - `$CMSE` a named data frame of CMSE scores for each model and each outcome, as well as model-mean CMSEs
        - `$experimental` contains the experimentally-determined effect of the intervention on posttest and outcomes
        - `$nonexperimental` contains the model-implied effect of the intervention on posttest and each outcome
}
\description{
`Causal Mean Squared Error (CMSE)` is a fit index that quantifies misfit 
between the model-implied causal effects of an intervention and the 
empirically determined effects.

CMSE: Causal Mean Squared Error index
}
\details{
This is quantified as the difference between the (regression-determined)
empirical effects of the intervention on distal outcomes, and the model's 
predictions about those distal effects given the empirical proximal effect.
See details for more.
}
