% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSE.R
\name{computeExperimental}
\alias{computeExperimental}
\title{Compute Experimental CMSE Elements}
\usage{
computeExperimental(
  dataset,
  intervention,
  posttest,
  outcomes,
  covariates = NULL
)
}
\arguments{
\item{dataset}{the actual data as a data frame or mxDataStatic object}

\item{intervention}{character string containing the name of the intervention column}

\item{posttest}{character string containing the name of the posttest outcome}

\item{outcomes}{character vector including names of all additional outcomes}

\item{covariates}{character vector including the names of all covariates to be included}
}
\value{
a list containing elements:
        - `$a` contains the experimentally-determined effect of the intervention on posttest
        - `$c` contains a named list of all outcomes; each outcome contains the experimentally-determined effect of the intervention on that outcome
}
\description{
Computes c.experimental and a.experimental values for the CMSE computation.
}
