% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSE.R
\name{computeNonExperimental}
\alias{computeNonExperimental}
\title{Compute Model-implied CMSE Elements}
\usage{
computeNonExperimental(ctl, models, posttest, outcomes)
}
\arguments{
\item{ctl}{the control group only of the experimental data set}

\item{models}{a list of mxModel objects (like those returned from makeARPanelModels) for fitting}

\item{posttest}{character string containing the name of the posttest true score (see details), or a vector of the same length as models containing the true scores for each}

\item{outcomes}{character vector including names of all additional outcomes, or a list the same length as models with outcomes for each model}
}
\value{
a list containing the model-implied effect of the posttest on each outcome
}
\description{
Computes b.nonexperimental and a.nonexperimental values for the CMSE computation.
}
